Broadcast updates to Game board (instead of refreshing)
Since we're already having to prepare the HTML and render it to the
user that makes the Cell Action (highlight, reveal, toggle flag, etc),
we can just reuse that prepared HTML and broadcast the appropriate
update to other players. Appropriateness is determined by whether or not
the Game has just ended.
- If Game is still current: update the :current_game_content frame
  - Note: In this commit, we update this frame name from `:current_game`
    (which was a bug) to `:current_game_content`.
- If Game has just ended: replace the :current_game frame

This eliminates the full-page refresh that was previously being
broadcasted to other players, which:
- save on 1 server request per connected player
  - which includes a re-render of the board. again, per connected player
... however, it also:
- Introduces the option to render an outdated version of the board
  during fast game play. We'll to fix this by implementing Game board
  versioning in a following commit.