Replace time-based Game#just_ended? w/ virtual attr
Now that we're broadcasting updates to the Game board (instead of
broadcasting a page-refresh) we no longer need a time-based
Game#just_ended? check. We need it before only because the page reload
had to determine if a Game just ended starting from a fresh request
context. Whereas now we can know within the same context. i.e. just by
setting a virtual attribute, which we can access within the same request
context as needed.
- Note: I tried using Game#ended_at_previously_changed?(from: nil), but
  this is very dependent upon the order of saves. We would have to make
  it so that the last save to the Game record during a Game-end event
  was the one that send Game#endd_at, but this is too big of an ask
  when we have a lot of things going on -- especially e.g. creating a
  GameEndTransaction to record the event, etc.

There wasn't previously any test examples for Game#just_ended? so I've
added those as well. That was just an oversight.