Implement versioned_replace for Game content
... using a custom Turbo JS helper:
Turbo.StreamActions.versioned_replace

Since we will have many players sending many new moves/updates, we need
to make sure we don't replace other players' Game boards with old /
outdated versions of the Game board.
- This implementation is highly copied from minesvshumanity. See:
  https://github.com/radanskoric/minesvshumanity/commit/dd14911c3fd256a43208f4bbae6ca7d62b1aa870

Note: This also switches us from doing an `update` to doing a `replace`
(except now: `versioned_replace`) for "Current" Game board updates so
that the `data-version` attribute, which we've placed on the
turbo_frame_tag, can update at the same time. We use this attribute to
determine whether or not a broadcasted update should actually be applied
or not.

Note: We use a "virtual" timestamp (`Time.now.to_i`) to represent the
Game#version since... it's an inexpensive way to do this. The first time
we call for Game#version we cache the value for consistency per Game
instance.... and beyond that, a timestamp is good enough to prevent
out-of-order updates from coming in.
- We don't have the luxury of just counting Game moves since we don't
  record Cell higlights / dehighlights in any transaction history (and
  it doesn't make sense to do so).
- And I also don't want to trouble the Database with touching
  Game#updated_at for every move made as this is "wasteful"...