Fix non-user-specific display of just-ended Game footer
Previously (before the last few commits), we simply refreshed the page
on Game-end. This allowed the current session + current User-specific
content to load. Including the signature form (for users that did
participate in the game), as well as the `👀` for "current user"
indicator in the Duty Roster. But, now that we're no longer refreshing
the page and, rather, broadcasting the just-ended Game container...
we've ran into an issue where the last user to make a move (the
Game-ending move) was having their User-specific content broadcast to
all other Users.

To fix this, we can still broadcast the Game-end state/container, for
the most part... but we need to separate out the User-specific content
and load that bit in on a per-User basis. In this commit, we accomplish
this with a turbo_frame_tag that is sourced/loaded on-load. The source
of the turbo_frame_tag will load a target that is then rightly current
User-specific, as is needed.

To make this work, we also need to remove the `game.just_ended?` check
when displaying the User signature line. This is no longer possible
within the current User-specific frame load context, but it also hasn't
even been needed ever since extracting this content to the "Just Ended
Game" namespace. Simply loading this partial from the "just_ended"
path/namespace is proof that Game#just_ended? is already true.

Note: To make the secondary loading of the sourced turbo_frame_tag feel
a bit less jarring, we show a loading spinner by default in this space.