Ensure idempotency of "Just Ended" Game frame replacement
I thought that broadcasts were for *other* sessions only -- and somehow
wouldn't be sent to the current/originating request/session. But I can
find no evidence that Turbo is supposed to work this way and... what I
was seeing was a doubling of the replacement of the
games/just_ended/container. This was the first time that was noticeable
because just-ended Game replacement includes a portion of code that
isn't within the turbo_frame_tag... so the second replacement added that
section of code twice. e.g. we were getting the
 #new_game_notification_container div rendered twice at Game-end.

To fix, we must make the update idempotent. This way we don't get
ourselves into a different DOM state for the originating session vs
other sessions. To make the "replace" idempotent, we just have to put
all of the changes inside of the turbo_frame_tag.

Note: I tried doing it the other way first. i.e. just don't render the
update and, instead, rely solely on the broadcast for updating Game
state (for all sessions). But, this breaks down as soon as the web
socket breaks down. Which seems to happen fairly quickly. And, in this
case, the users sees an unresponsive Game board. To fix, he/she must
then refresh the page (to reconnect to the web socket). So, the approach
in this commit is better since it will always ensure the originating
request goes through and, in turn, this will also reconnect the web
socket, ensuring smooth Game play/state from there on (again).