Wait for originator exit to send New Game Notification
This fixes an issue where the "New fleet deployed. Reload ↻ to join in!"
notification would (momentarily) show up even for the user that clicked
on the New Game creation button. This was happening because broadcasts
go to everyone, even the originator session for the broadcast. So,
because the notification was being prepared on the server and then
simply passed on / delivered in real time, we would see the notification
appear for just a split second before the page changed over to the new
Current Game view.

Since there's no way to tell Turbo -- more accurately ActionCable --
which sessions to exclude from receiving a broadcast, the simplest fix
I could come up with here was to simply delay the creation and send of
the notification by just enough time to get off the page and on to the
next. This way, the notification doesn't flash into the DOM for split
second first. Plus, it's good anyway as it gets the web server's
processing time off of the creation of the notification and back to just
serving web requests.