Change design of `inline-edit` form/controller
Instead of building up a staged toggle state for display of inline
forms, we can just "Cancel" the form by including a hidden "Cancel" link
that calls the "show" action via a turbo_stream (update).

I think this is more inline with expectations for inline editing of
Turbo forms and involves less Stimulus.js code/magic.

Plus, we now have an easy-use "Cancel" link (really: "🙈" link) next to
the inline form. This will help people that don't otherwise think to use
keyboard commands (which is probably most people).

Next, we'll need to do this for all the other inline-edit forms.

References:
https://www.driftingruby.com/episodes/inline-editing-with-turbo
(Note: This is a paid service screencast.)