Don't transpose mobile Game boards anymore
We can't try to transpose the Game board for mobile anymore because we
recently stopped telling each subscriber to go refresh their own page.
Instead, we're now pre-packaging up the new Game board state and
broadcasting an update with that HTML content to all stream subscribers.
Which means that if a mobile browser made the move then the Game board
would transpose itself for mobile and then send that HTML out to all,
including non-mobile sessions.

Removing the transposition behaviors also reduces the number of caches
needed, as we no longer need to potentially cache mobile + non-mobile
versions of every Game board. (Especially experienced while browsing
Past Games).

And finally, removing transposition also reduces a few of the more
complicated parts of the code. So this is an overall win, I think. It
actually doesn't feel too bad to have to scroll horizontally in mobile
because, in fact, then you only have to scroll in one direction. i.e.
in landscape, the vertical height of the game board does fit within the
screen height. While the opposite is not true: in portrait, the
horizontal width of the game board does not fit within the screen width.
- There's just one annoyance left... when the board is refreshed on
  mobile, the horizontal scroll position of the Game board is reset. So,
  if you were viewing the right side of the board then a refresh will
  pop you over to the left side of the board and you have to scroll
  over to the right side of the board again to see where you were.

Note: This CSS-only solution almost worked, but it wraps poorly on
mobile and otherwise causes a flash of unstyled content on page load:

```CSS
table.transpose {
    @apply whitespace-nowrap;

    tr {
      @apply block float-left;

      th,
      td {
        @apply block;
      }

      & > *:not(:first-child) {
        @apply border-t-0;
      }
      &:not(:first-child) > * {
        @apply border-l-0;
      }
    }
  }
```