Show # players at ⚓️ when Standing By
Previously, we only showed the count of players in the Game session once
the Game had actually started. But there's no (good) reason, that I can
recall anyway, not to just always show the # players in the room. This
gives you a sense of how many people are there that could potentially
start the game, which is kind of cool.