Fix board header content highlight/selection on drag
This didn't happen in Firefox, only Safari that I know of, but...
If a User clicks down and drags their mouse just a little bit over a
revealed Cell number on the Game board, then it could cause a
highlight/selection to appear over the table content header. It's just
some weird glitch but... we can make it not happen by apply the
`select-none` CSS class/rule more liberally: to the entire Game board
and surrounding content. Then, we can selectively add back in
`select-auto` behavior as needed.

Note: I still didn't add `select-auto` back into the Current Game
"Status" text because if I do then *that* text will now be awkwardly
highlighted/selected on mouse drag across a revealed Cell. I would
rather have had that text be selectable but since this behavior is so
easily reproduced (unwittingly), I'm willing to sacrifice this text
selection. We could, however, perhaps only apply this to Safari if the
need arises? I'm fine with the Flags-to-Mines remaining details + the
Game Timer not being selectable text on the Current Game page, however
we can make them use `select-all` on the Paste Games page without
problem.

Also: Simplify the display of the Flags-to_Mines remaining details text.
Both for simpler rendering, but also to make this text actually
selectable (on the Past Game Show page) as a single line of text. The
previous ul/li version created a stacked, multi-line version of the
copied text.