Add long-press as "right click" event support for mobile
This is perhaps still a bit experimental at this phase, but it seems to
work for the most part. I've tested so far just on mobile Safari on iOS.
But... now we can long press (for 250ms+) to toggle a flag. Since one
can't see through their finger, we add a "ok it happened" indicator by
animating the flags-to-mines ratio text, which should generally be in
view on mobile. Also, if the mobile browser supports it (which Safari
does not), we attempt to inspire a haptic engine vibration as well.

Because cell higlights don't really seem to play well with mobile, I'm
just disabling them for mobile browsers for now. What I'm seeing that
made me do this was the cell highlight effect not consistently applying,
and then, often, sticking around longer than it should. Plus, the effect
doesn't appear on click+hold, it would only appear on tap+release. So
there's likely interference there with the mobile browser's long-press
functionality. But, cell highlighting isn't critical functionality and
can be skipped over on mobile. It's much more important to actually be
playable.