Update messaging about Mobile support
Given the updates in the previous commit to allow for long-press -> Flag
in mobile...

Remove the "FAQ" entry on the About page. Nothing special need be said
anymore about Mobile. Especially since...

Update the "Rules of Engagement" to offer action descriptors that are
inline with the current browsing experience. i.e.
- "Click" vs "Tap"
- "Right Click" vs "Long Press"

To support this, I chose to refactor towards passing in our
Application::Layout instance (accessed via `layout` in views) (instead of
the view template (accessed via `self` in views)...) for optimally
answering the `mobile?` question.