Switch to preferring `rails dev:cache` for cache toggling
Rails has a built in tool for toggling caching in development:
`bin/rails dev:cache`. This toggles the config options in
config/environments/development.rb by creating/removing a tmp file:
"tmp/caching-dev.txt"

This is great as it gives us an additional degree of freedom for
testing/development over just tying caching to the `App.debug?` flag.

To make this toggle state more clear in development, add icons to the
_dev_flags.html.erb partial to indicate when caching is enabled or not.