Hide scrollbars on the Game board
This mainly affects mobile, but will also effect desktop if the browser
window is skinnier than the Game board. The scrollbars don't really add
any useful information... and in fact, they kind of get in the way.
- Especially after implementing auto-restore of horizontal scroll
  position on turbo stream replace (coming in the next commit).

Tailwindcss v3 doesn't have a utility for hiding scrollbars, so we roll
our own utility. Reference:
https://www.subframe.com/blog/hide-your-scrollbars-but-still-scroll-with-tailwind-css