Keep horizontal scroll position on Game board updates
Previously, the turbo_stream replace on the Game board content (which
happens on every move made -- either by the user or by other players)
would wipe out the horizontal scroll position of the Gameboard. This
effects mobile moreso, but also applies to desktop if the browser window
is just too skinny for the Game board.

To fix this, we add a scroll_position_controller.js...
... which is responsible for saving the current horizontal/vertical
scroll position for `this.element` (in this case: the Game board) and
then restoring it after replacement via turbo stream.

This new behavior must apply to both the current Game board as well as
just-ended Game boards. Else, the game-ending move will wipe out the
scroll position, which is pretty jarring (as was the original issue).
- So, to satisfy this, we have to store the Game board's scroll position
  into an element that's outside of the turbo stream replacement area
  for the Game-end event. The ideal spot for this seems to be the one
  view template that is common to both current and just-ended Games:
  app/views/home/show.html.erb