Always send Game board updates as "versioned replace"
This will hopefully fix some wonkyness I saw on production where
sometimes if a Game ended and I quickly clicked to start a new Game I
would see the new Game view (for a split second) but then it was almost
instantly replaced with the just-ended Game view from before the
"New game" click. And then, right after that instant, I'd see the
"New fleet deployed." notification message. Yet I was the only player in
the channel so... there has to have been some sort of race condition
going on. By using versioned replace we should be able to side-step this
since the version number will always be equivalent between both the
rendered version of the Game board and the turbo stream-broadcasted
version of the Game board (since they're using the same HTML string).

To solve this, we can:
1. Just always call WarRoomChannel.broadcast_replace
3. Make WarRoomChannel.broadcast_replace just always do a versioned
   replace.
I like this because it abstracts away the fact that we're ever doing a
versioned replace, making it just an implementation of the
WarRoomChannel object.

To support this, we must add the current game version to both the
Current Game view's main container frame as well as the the Just Ended
Game view's main container frame, so that our versioned replace JS
function has something to go on for whether or not to replace the
existing Game container.
- This should be the fix for the specific problem first described in
  this commit message.