Fix: Don't show "Custom" btn on Game-end to non-signers
Previously, if the Game-ending move was performed by a signer (a User
that has signed their name), then we would default to rendering the
"Custom" button in the New Game Content partial. This would then show
up for all users (even non-signers) since we're now sending the
generated HTML to all users at Game-end. And the opposite was true too:
If the Game-ending move was performed by a non-signer, then we would
default to rendering the New Game Content partial without the "Custom"
button (even for signers).

To fix, we:
1. Default to always showing the "Custom" button, regardless of who
   made the Game-ending move
2. Use JS to quickly remove the "Custom" button on-connect to the
   Stimulus JS controller for non-Signers

This approach is imperceptible and doesn't require another request to
solve. I mention this because my first attempt at a fix for this was
just that: Eager load a turbo frame for each user. This wasn't good,
though, because this meant an additional request per each connected
User, plus there was a very visible gap of time where the buttons were
either not rendered yet (showing a Spinner instead) or some other...
trick. But, regardless the gap of time + additional server request were
obvious disadvantages over the solution put forth now, in this commit.

Note: I don't love that we're kind of just hard-coding DOM ID strings
as requirements/expectations for this JS to work... but, I'm trying this
out because: I've made a distinction with this Stimulus JS controller,
in particular, by not making it generic. Instead, its path matches that
of the view that is relying upon it. So there is a more strict
connection between the 2. Plus, any attempts at DRYing up the DOM ID
Strings would add further complexity while littering up some other,
mostly disinterested object. i.e. instead of muddying up e.g. the View
view model with DOM ID constants or something... I want to try just
"knowing" that DOM IDs are unique things that, if changed, should first
be searched on for matching changes elsewhere. I think this is generally
acceptable. So, instead of forcing a single place to change, we can just
rely on search/replace. A bit like this same philosophy with
TailwindCSS.