Add UIPortal::UnsupportedBrowserTestController
We're getting an error when Rails' `allow_browser` macro blocks a
browser. See:
https://app.honeybadger.io/projects/129535/faults/116131377

To be able to test this condition ourselves, we add a special UIPortal
controller that will not allow any browsers.

The actual bug fix is yet to come.

With this change, we do indeed see the same error message as reported by
Honeybadger:

> File /Users/paul/dev/minesweeper_alliance/public/406-unsupported-browser.html does not exist