Make `allow_browser` work w/ dynamic Error Pages
The `allow_browser` macro expects a file to exist at /public/404.html.
Without this, it raises an error:

```
File /Users/paul/dev/minesweeper_alliance/public/406-unsupported-browser.html does not exist
```

See: https://app.honeybadger.io/projects/129535/faults/116131377

To fix this, we need to:
1. Add a `block:` attribute/action that specifies a browser redirect vs
   just rendering a public html page.
2. Exclude our custom Error Pages controller from the `allow_browser`
   before_action. Else, we just end up in the same spot when the
   custom error page blocks the request for that browser again.

Finally, we repeat the new `block:` attribute/action in the UI Portal
test page link.

The approach taken in this commit requires us to remember to include the
AllowBrowserBehaviors mix-in for all controllers that should enforce
browser version. This is both good and bad. An alternative approach
would be to create a UserPortal or PlayerPortal (or something)
that "all" (most) controllers would inherit from instead. Then, a
core-level controller like ErrorsController could inherit from
ApplicationController while being outside of the purview of the
`allow_browser` macro / redirects.
- Problems with this: There may still be some controllers that are
  legitimately in the User/PlayerPortal but still shouldn't adhere
  to the browser redirect (e.g. the About page).