Fix erratic flag toggle via ctrl + left click in Safari
Safari triggers more events/callbacks when ctrl + Left Clicking than
Firefox does. So I didn't notice this earlier, but...

Using `ctrl` + Left Click on macOS can trigger the `reveal` event
handler as well as the `toggleFlag` method. But if we let both go
through then the flag will generally be set and then removed just as
quickly. So we need to just return early on the `reveal` event handler
and let toggleFlag() do the work.

Also, we add actsAsLeftClick() to lib/mouse.js, to help differentiate
Left Clicks that were actually meant to be Right Clicks (e.g. when the
`ctrl` key was held down and the OS is macOS). This allows for proper
event filtering in Safari as well.