Sort Users by first move (DateTime) in Duty Roster
Previously, we were just getting Users sorted by User#id, I guess. I
don't really understand why PostgreSQL needs the things it needs...

Anyway, by starting from the User model (instead of the Game model which
probably never made much sense), I was able to find a way to group Users
by earliest CellTransaction DateTime (a.k.a. when they "joined_in_at").
This is much more useful than before, as it then properly reflects
participation order in the Duty Roster. i.e. the first-listed User was
the first to make a move, the second-listed User was the second to make
a move, and so on.