Fix unwanted turbo frame nesting on signature form
Previously, canceling the user signature form (after tapping to sign
one's name on Game end) would update the `signature_user_<user_token>`
turbo_frame with a copy of itself. So we'd end up with e.g.

```
<turbo-frame id="signature_user_18c81778-f870-474a-aaa9-805535e49edd">
  <turbo-frame id="signature_user_18c81778-f870-474a-aaa9-805535e49edd">
    # ...
  </turbo-frame>
</turbo-frame>
```

This was because we were performing an `update` action on turbo_frame
that rendered a partial with the same turbo_frame in it.

To fix, we could have perhaps changed to a `replace` instead of an
`update` but I prefer `update`, by default, since it's changing less in
the DOM.

Also, I had mistakenly added the same `signature_user_<user_token>` DOM
ID to the signature form itself, which was compounding the problem.
It works well without that DOM ID being inserted into the form, so we
remove it as well.

Finally, we can remove
app/views/games/participants/show.turbo_stream.erb in favor of just
letting the form render app/views/games/participants/show.html.erb,
which already has the `signature_user_<user_token>` turbo_frame on it,
so Turbo will know what content to grab/replace. This simplifies the
total stack, even if it's a little less efficient.