Fix signature input field width on mobile
Previously, the signature input field was too wide. It was being forced
to expand past the margins of the mobile browser view due to the `w-80`
class.

The fix was to change the form input field to take up 100% of the
natural width for mobile views. However, to achieve this, we had to make
the container turbo_frame be a full-width element as well because it was
otherwise constraining the input field to its unnaturally-constrained
width.