Add manual/programmatic Honeybadger notification ability
When trapping on an exception, we can't just swallow it and move on...
or we'll never really know that the event occurred. So it's generally
best to at least log the exception. But, since we now have Honeybadger,
we can do one better and actually ship exceptions details on up to
Honeybadger for our information. This can cause a lof of noise
sometimes, so we may need to adjust... but it's a good way to get us
doing the right things in our code. e.g. reducing the scope of
exception handling blocks, etc.

Of focus in this commit, is our current catch-all for any StandardError
(especially: Games::Current::Board::Cells::ActionBehaviors::Error)
when doing any board "action" (Reveal, Flag, etc). I previously added
this rescue block to handle, presumable "lost connections". i.e. if the
Game had ended and we then came back to a stale War Room state and tried
to make a move. This would result in a Cell not Found error (because
we scope Cell#find to the current Game). So my hope is that I can learn
some more details about these situations over time.

To accomplish all of this, I'm adding some fairly generic classes that
I've started in other projects and that should be able to carry over
into projects in the future as well as I improve them. They're in `lib`
for this version reason.