Auto-reload page on new Game for spectators (vs participants)
It's become clear that we need to fully separate out spectator-only
content/behavior from participant content/behaviors for War Room games.
Basically: so we don't spend time/resources rendering (or deciding
whether or not to render over and over again) content/features/behavior
that is only relevant to participants vs spectators of the Game. e.g.

For Participants:
- Disconnect from WarRoomChannel (:current_game) on Game-end so that
  the user has time to sign their name, review Game-end stats, etc.
- Show the "New fleet deployed. Reload to join in!" alert on new Game
  start.

For Spectators:
- Don't disconnect from WarRoomChannel (:current_game) on Game-end for
  spectators. This way they will continue to follow along with new
  Games, uninterrupted.
  - There's no need to opt in to new Games because spectators can't sign
    their name anyway.
- Don't show the "New fleet deployed. Reload to join in!" alert on new
  Game start. This allows for casual, passive observation over multiples
  games, which is nice.

To achieve this, I pushed the decision point up the stack as far as
possible, which ended up being the Controller (again). So,
Games::JustEnded::FooterController is responsible for choosing which
footer to render.
- We choose the footer because this is where all the differing content
 is. The Game board, etc, is the same for all users.