Show "Fleet Connection Status" dot next to War Room link
Gray = disconnected from the WarRoom Channel's Turbo Stream
Green = Connected to the WarRoom Channel's Turbo Stream

This uses a JS MutationObserver
(https://www.javascripttutorial.net/javascript-dom/javascript-mutationobserver/)
to monitor the turbo-cable-stream-source#current_game_turbo_stream DOM
element and respond to the presence of its `connected` attribute. When
present, we're connected. I'm not sure how to tap directly into the
stream that does this... but this is good enough for us (if not the only
way).

New Phenomenon:
- It is now more apparent when we temporarily lose and re-establish
  connection to the War Room channel. e.g. when starting a new Game,
  the whole page refreshes--which re-establishes the connection. This
  shows up as the connection status light switching from (quickly):
  green -> gray -> green. So... may want to do something about this to
  maintain the integrity of the stream connection in the future.