Mass refactor: Participants -> ActiveParticipants
... and Spectators -> Observers.

I'm trying to clarify organization of views, View Models, etc in light
of the upcoming "Observers" list/distinction. "Participants" was
too vague... is it an active participant (a player that made a move in a
game) or a passive one (just an observer of the game).

Similarly confusing is the dual-role of our current DutyRoster. The
DutyRoster (cache) keeps track of all participants (active and passive).
It's more of a potential list of active Users. I've left further
clarification here to a future commit, though some cursory steps were
taken by renaming non-cache-related "DutyRoster" objects to just
"Roster".

Finally, it continued to feel weird thinking of "Participants" as having
an edit form that controlled their signature. Both because of the
active/passive thing, but also because the model behind User signature
editing is User. Also, in fact, this is an action only ever on the
"Current User" (concept). So I've renamed these
objects/controllers/partials as well, from e.g.
Games::JustEnded::Participants::Edit ->
Games::JustEnded::ActiveParticipants::CurrentUser::Edit.

Functionality has not changed in any of this... it just paves the way
for new functionality: passive participants / Observers.