Refactor DutyRoster -> FleetTracker
FleetTracker represents the number of participants (Users)--both
active ("Active Participants") and passive ("Observers")--currently in
the War Room. This count is reset at the start of each new Game.

Motivation for the name change was to distance it from the UI concept
of the "Duty Roster" -- which represents only the Active Participants
for the Game. FleetTracker, however, keeps track of Observers as
well, and is only meant to serve as a high-level "headcount" of people
in the WarRoom during a Game.

Along with refactoring the class name, I've also refactored:
- DutyRoster.participants -> FleetTracker.registry
- DutyRoster.cleanup -> FleetTracker.prune
- DutyRoster.clear -> FleetTracker.reset

There are no functional changes herein.