Fix missing flag in "Mine Flag" Pattern
We had a bug in the seed file that was duplicating one of the
coordinates, probably due to a formatting adjustments error. I think I
just ended up accidentally deleting the `3` in `[23, 22]`, which caused
a duplicate entry for `[2, 22]`, which meant we had a missing flag in
the right-most wave of the Pattern.

Also: Remove the flag from the bottom-right corner of the buoy... just
b/c that seems better to me today.