Add Games::Current::Roster View + broadcasted updates
This is a big step towards displaying the Current Fleet Roster... though
it's not done yet. It's hard to fit this into the UI, so I think I'm
going to "punt" and make it into a tooltip. But, our current tooltips
implementation doesn't allow for complex HMTL like this as it's a pure
CSS solution. So, for now, I'm just hiding the roster unless App.debug?.

Refactor BroadcastCurrentFleetSizeJob ->
Games::Current::BroadcastRosterUpdateJob, and have it handle Fleet Count
update as well as Roster update, to minimize background job processing.
They're both the same topic anyway.

Move handling of fleet size updates from Games::Current::Container to
Games::Current::Status, since the fleet count is actually in the
"status" area of the Game container. There was no need to push it out
to the Container. I think, at the time, I was just thinking of
"container" as basically the "Game Show" page... but still.

Stop trying to limit Fleet.count at `1`. This is a vestige of the days
when our tracker was doing pure increment/decrement without user tokens
for accountability. I think we can trust it now, and if not then I want
to see that show up in the UI at some point. Seems fine for now and it's
worth removing the incremental complexity.