Remove the "Previous Op" Page Nav from War Room view
I don't think it's really needed anymore. Now that we have the
"disconnect from the :current_game stream on Game end" logic with the
"Reload" link... I don't think there's the same risk of being confused
about where the Game just were playing went anymore. Which is why I
had originally added it to this view.

So... we can just de-clutter the view a bit now.