Add new Tooltip JS Controller + CSS (via Floating-UI)
Our existing tooltips solution is lacking. It's a little ugly, for
one... but it also bunts on dynamic positioning based on screen
position. For that, we need to introduce JS. For that, I've selected
Floating-UI.

Floating-UI provides the JS component that drives all of this. See:
https://floating-ui.com/docs/getting-started
- It appears to be the successor to popper.js (which is ubiquitous).

This commit sets out a sensible set of defaults for configuring our
Floating-UI JS logic/config, along with CSS for our upcoming tooltip
components (views)--see next commit.