Increase font size on all cells while removing padding
This makes the numbers slightly more visible while sacrificing nothing
as far as overall Game board width/height (it stays exactly the same).

I also toyed with reducing the size of the Cells, but decided against
it as it does make the Game harder to see/play. minesweeper.online uses
what would be the equivalent of size-6, which is quite a bit smaller.

We could offer a "zoom" level option in the future (as they do)...