Slide In/Out Roster from top-right edge of <main> container
Click on the Fleet Count indicator "button" to toggle display of the
Roster (of all Users present in the War Room).
- We use el-transition.js to turn this into a slide-in/out effect using
  the required copious data-transition-* attributes and by re-using
  menu_controller.js.

To support this change from the fleet count being in a `<small>` tag
to being in a `<button>` tag (as this is the right thing to do for
semantics), add a `.small` CSS utility and define it to match the
default tailwindcss style for `<small>` tags. Reference:
https://unpkg.com/tailwindcss@3.4.17/src/css/preflight.css

Still TODO:
- Remove games/current/status partial from games/current/controller
  - And then somehow update Game Status on transition from
    Standing By -> Sweep In Progress