Add optional externalClose target to menu_controller.js
This is useful for our sliding menu b/c it gives a clear way to close
the menu. And, in our case with the Roster, it's the only way to close
it on mobile (apart from just refreshing the page).

We give the externalCloseTarget enter/leave effects (via
el-transition.js) to enhance the effect.