Don't re-render Fleet Roster on Game start
If a user opens the Fleet Roster tray, we don't want to lose its state
when the Game starts (i.e. on first Cell reveal). Doing so is a bit
surprising UX.

To fix this, we have to move the rendering of the Game status outside of
the game "content" partial/div, as this is what's being
replaced/refreshed on each Cell action (reveal, flag, etc).

Next, we have to then refresh the Game Status + Emojis manually on
Game start.

This commit includes some refactoring to create a Game::Current and
Game::Past View Model. These are meant to DRY up some of the repeated
needs of their respective View Models. The Games::Current refactor was
needed in order to make the rework sensible... the Games::Past refactor
was for symmetry/completeness.