Don't fail if no current game when broadcasting Roster update
Previously, Games::Current::BroadcastRosterUpdateJob would fail and
raise an unhandled exception if there was no current Game:

  ActionView::Template::Error: undefined method 'status_standing_by?' for nil

The above was found by a Honeybadger alert in production.

We fix this by just getting rid of the requirement to pass in a current
Game. The only thing we were using the current Game for was to render
a specialized Emoji set based on the current Game status... but this
doesn't seem worth the headache. The roster just shouldn't be tied to
a Current Game anyway, as it's more of a "War Room" feature than a
Current Game feature. -- We'll start to build this out more in the
near future.
- The one drawback of this is that we always show ⛴️ ⚓️ as the "title"
  of the roster panel.