Allow view of roster on all pages w/in Home / War Room
Previously, we had attached the roster to the Current Game (status).
This didn't allow for viewing the roster during other War Room states:
- At end of game
- When creating a New game

Now all of these states are possible, just by moving the rendering up
the stack.

To support this, we untie the roster from the Current Game state
(since there isn't always going to be a Current Game). We can tie this
back in later if it still seems important.

Still TODO:
- Don't rerender/close the roster when creating got to New Game screen
- Don't rerender/close the roster when creating a new Game