Persist open/closed state of Roster slide menu
When opening the new Roster slide menu, save that fact into a browser
cookie. When closing the new Roster slide menu, delete the browser
cookie.

Then, on future page loads, render the slide menu as in its "open" vs
"closed" state based on the presence (or not) of this cookie.

This is a useful feature for persisting state across site navigation...
but for now it also serves as a band-aid to our current implementation
of the War Room. i.e. when we create a new game, our current method of
refreshing the Game board / view is to just refresh the whole page (vs
only refreshing the needed components of the page). My goal is to fix
this in the future so that we never see the flash of "unstyled content"
were the Roster slide menu appears to fidget between a closed/open
state on browser navigation/refresh/etc. This fidet state comes from the
Turbo cache + refresh of the last-accessed page state.

Also: Code cleanup/refactoring in SlideMenu