Extract CollapseContainer from Games::Current::Rules
This is just a refactoring on the old "Rules of Engagement" Collapse
Container, to get it up to par with our new Slide Menu code.

Extract Games::Current::Rules::CollapseContainer out of
Games::Current::Rules so that it's its own thing, and so we can name
the view model (CollapseContainer) after the partial
(_collapse_container.html.erb).

Also: Don't keep hold of the current view context any longer than
necessary. And prefer keyword argument-style passing of the `context`
parameter.