Show Observers on Just-Ended and Past Games
Observers are passive participants. They've joined a Game but haven't
made any moves in the Game, and certainly haven't, started, or ended
a Game. I'm still on the fence about if a User is an observer if they've
created the Game... but currently the code excludes these as well. I
think that's the right choice. Until you've revealed or flagged a Cell,
you haven't actively participated.

These queries are not simple... but they don't seem to be expensive.
Regardless, I am pondering adding a GameParticipationTransaction to
simplify the long trail of joins via, currently:
CellTransaction -> Cell -> Board -> Game.