Migrate users to signed user token cookies on visit
This was an oversight... I forgot to come back to cookie storage of
User tokens before going live to production to make them signed. (It was
easier to debug things when I could read the user token cookies in plain
text.)

Basically, a signed cookie makes it tamper proof. So people can't copy
a User's UUID and place it in their own `user_token` cookie to take
over someone else's identify on the site.